feat: Add mutation display name, and mutant diagnostics
Add a configurable display name for mutations, along with a diagnostic message, used when a mutant is undetected, for mutants. These are baked into the generated program and are output by the test harness.

The location of each mutation is stored and used to create diagnostics and as part of other miscellaneous output.

All possible diagnostics are generated at compile-time and baked into the generated program for use by the test harness. The diagnostics are stored using an emitter which writes to a string buffer, instead of stderr.