fix: Use ids instead of pointer equality to dispatch substitutions
Pointer equality was never guaranteed when referencing mutations through the active mutant handle. This caused substitutions to not apply correctly. To fix this, we now bake a u32 id into the mutation metadata which can be compared against at runtime, even if the data gets moved to a different location by the compiler.