feat(range_limit_swap): Add mutation operator
Swap the limits (inclusivity) of range expressions.

Simply changing the limits on the syntax level does not work, because the desugared types are different for each kind of range expression. Instead, we increment/decrement the end bound of the range to emulate the same behaviour for the common cases (integers), where this is possible.