feat: Formalize substitution conflicts
When batching mutations, we avoid combining mutations with conflicting substitutions. This used to be a simple location comparison, but this outrules some combinations of substitutions which are valid (e.g. local binding substitutions never conflict with each other because of how they are expanded).

Note that this does not concern the compatibility of mutations (whether they have any effects on each other) only that their corresponding substitutions can be applied at the same time.

This also includes a fix for a subtle batching bug which was caused by a loop break (instead of a continue). Our `continue_break_swap` mutation operator would have caught this one. :^)