feat: Collect target reachability from tests to avoid batching conflicts
For each mutation target, we now collect all the tests they are reachable from (including the respective distance). This is used to avoid batching mutations together which are reachable from the same test, to avoid having mutations conflict with each other. This will be useful for evaluation; it could be used to only run tests which reach the current mutations.

The above guarantee could also be used to disambiguate test failures. It makes it possible to determine which mutation caused a test failure, since only at most one mutation is applied per test case.