build: Update Rust toolchain to nightly-2022-06-13
Update Rust toolchain and dependencies to their latest versions.

Since nightly-2022-05-28, the Rust toolchain ships with multiple versions of smallvec, which cannot be resolved in the mutest-operators crate. The compiler-supplied version of the library continues to resolve correctly in mutest-emit, since it links against the compiler directly. To resolve this, the right version of smallvec is now re-exported from mutest-emit, for use in mutest-operators.