feat: Implement Cargo subcommand
Add the `cargo mutest` subcommand and corresponding mutest-driver rustc wrapper binary. The binaries are modeled after clippy's cargo-clippy and clippy-driver binaries.

The driver is a rustc wrapper binary which is invoked by Cargo in place of the rustc compiler. The correct Cargo commands for this are invoked by cargo-mutest.

By reusing Cargo's build commands in this way, we ensure that the build process for the package/workspace is unchanged from a regular rustc build. This means that we also benefit from build caching, pipelining and even incremental compilation.