feat: Resolve the AST of ambiguous definitions (e.g. impls)
Certain definitions are not referred to by a unique identifier in the compiler. These were previously not resolved due to a lack of clear invariant which could be relied upon.

Definitions in the compiler use a disambiguator value to resolve ambiguities in a scope across definitions without a unique identifier. While this value is an incremented counter, unfortunately it does not always follow definition order as items from macro expansions are turned into definitions in a separate pass.

We define our own definition disambiguators, similar to the compiler's own, but with the values being assigned in definition order. This way, the disambiguators can be used to index the scope's items of the corresponding type directly. This works for the expanded AST as well, as it is guaranteed that items in the expanded AST and the HIR follow the same definition order.