feat: Add mutation span labels which can differ from the display name
The full display name of mutations is quite verbose, since it's meant to be used without any additional context, other than location. In diagnostic messages however, we print the corresponding code snippet, making some of the mutation descriptions overly verbose when used as a label as part of these diagnostics. With this change, mutations have the option of providing a separate, less verbose span label in addition to their full display name for this purpose.