feat: Register custom attribute namespace
In order to introduce custom, mutest specific attributes, we have to register a tool attribute namespace using the experimental register_tool crate attribute. We inject this into the code before macro expansion, which is the step that performs the relevant validation.

In addition, we now build with the mutest cfg configuration flag, which can be used to add mutest-specific attributes which get ignored by the regular compiler.