feat: Add #[mutest::skip] attribute
Add #[mutest::skip] function attribute, which can be used to specify that the function's body must not be mutated. A function annotated with this attribute will still have its callees searched. For example, this is useful for indicating that a non-test function is used as part of the test suite.