feat: Perform a profiled reference test run and various test orderings
We now perform a profiled reference test run of the unmodified test suite to sort tests by their execution time. This is based on the expectation that in most cases, mutations will not cause a major change in the runtime of individual tests. In addition, tests can also be ordered by mutation distance. Regardless of the ordering used, tests are placed in a queue by cycling through mutations, queueing the frontmost test for each mutation, until all tests have been reordered. This should improve the runner's ability to parallelize the test runs and the harness's ability to prune as many remaining tests as possible.

Based on the profiled reference test run, an automatic test timeout is also deduced, which may be overridden.

The final report now specifies the number of mutations which timed out or crashed their isolated process.