feat: Resolve externs of mutest-runtime when compiling generated code
When building mutest-runtime, Cargo passes `--extern` arguments mapping each dependency to its generated rlib file. When a crate is built against mutest-runtime, these externs have to be propagated to resolve transitive dependencies.

To solve this, we first bake Cargo's `--extern` arguments into mutest-runtime when building the crate, using a build script. Then, when compiling the generated binaries, we read the externs from the linked crate and propagate them to the current binary crate.

The `--extern` arguments Cargo uses to build mutest-runtime are extracted from Cargo's build plan, a JSON representation of Cargo's build steps for the workspace. The build plan parsing is largely based on the abandoned build-plan crate, with some very slight modifications to account for changes in the data format and an up-to-date serde version. The extern paths are stripped and are turned into relative paths so that a custom search path can be specified for them. The externs are separated by 0x1f (ASCII Unit Separator), similar to Cargo's own `CARGO_ENCODED_RUSTFLAGS`.