feat: Add options for printing type paths, use visible paths in codegen
We used to print fully-qualified paths for all definitions, but this caused issues with re-exported items declared in inaccessible modules. Now, instead of using the normalized definition path, we make use of the same logic that is used by the compiler in diagnostics to print and suggest paths to items that would be importable from external modules and crates. The implementation for this is heavily based on `rustc_middle::ty::print::pretty::PrettyPrinter::try_print_visible_def_path` and has been extracted into a more general form found in our `PrinterExt` trait.

In addition to the core logic described above, we also add an option for trimming the virtual qualifier prefix of paths pointing to scoped items, which are inaccessible from outside their declaring scope. Before, these items would generate fully-qualified paths which are only valid according to the compiler's internal reporting but not in actual code. Since these items can only be referred to from their declaring scope, an option to remove the virtual qualifier prefix was added, which should be an acceptable, valid representation for most use cases.