feat: Revert non-local macro expansions in generated crate code
During codegen and throughout analysis, we operate over the fully-expanded AST. This becomes problematic with macro hygiene, since the printed AST loses all hygiene information, causing incompatibilities between the crate's own code and code coming from external macros. To alleviate this, we attemp to revert non-local macro expansions, starting with derive and attribute macros.

As part of this (ongoing) initiative, we implement a simplified version of the compiler's macro expansion engine to perform module loading the same way the compiler does. This is needed to fetch the original, unexpanded items later. The compiler's own module loading code is deeply embedded in its macro expansion engine, so we unfortunately cannot reuse most of it.