feat: Mark targets which are called by unsafe code as tainted
Unsafe callers of a safe function likely expect not just a safe, but correct implementation of their functionality. To preserve the integrity of unsafe code, such functions must also be treated as being part of the unsafe scope. This unsafe scope extends to the entire subtree of callees.

As part of this work, the unsafety system has been refined significantly. Safe functions called from unsafe code are marked as tainted with unsafety by an unsafe block. In addition, each unsafety level has a source clarifying whether the unsafety comes directly from an unsafe block or the enclosing scope(s) of a dependent unsafe block. This allows us to model all dependencies of unsafe code; cases of unsafety seeping into otherwise safe code. :^)