feat: Implement custom path-based Debug representation for Test
Implement a custom Debug representation for Test which is shorter by only displaying the test path. The derived implementation was quite useless since it would end up printing two fairly large ASTs. Since we use the Test struct as a map key in a couple of places, it makes sense to provide a short, readable Debug representation highlighting the unique path of the test.