feat: Reverse test case expansion in generated code
During macro expansion, the compiler expands the `test` attributes into a descriptor struct item and an `extern crate test` declaration each. This results in code that has internal structs and duplicate extern crate declarations in the same scope, which are all invalid. To solve this, we reverse the expansion of `test` attributes with just the original attribute, letting the second compiler pass do its own internal expansion.