feat: Sort mutations by reverse "conflictingness" before batching
We introduce a mutation conflict heuristic into batching, which is the number of conflicts a mutation has with any other generated mutation. This heuristic is used to reverse sort the mutations before batching. This way, we effectively consider the most conflicting mutations first, and the least conflicting mutations last for batching. This seems to result in signigicantly better batches all-round.