feat: Update to rustc nightly-2023-09-26
mutest-rs was sitting on version nightly-2022-06-13 for over a year (one year and three months to be precise :')). A lot has changed in the compiler since then and a lot has not, but in general, the code has become much cleaner from the compiler API changes.

It is exciting to be back! :-)