feat: Remove locking from the active mutant handle
We previously used an RwLock to avoid unsafe code, but this came with a large performance penalty. Since (ideally) no other thread runs when we modify the handle, the behaviour stays safe and the same. To ensure that this guarantee stays, the function to modify the handle is made unsafe and crate-private. As for lingering test cases from previous test runs, their behaviour should remain the same, regardless of whether the handle performs locking or not.

This was primarily done to remove the potential for deadlocks in cases where a mutation was placed in a hot path and the corresponding test times out and remains running in the background as a lingering thread. See the documentation for `std::sync::RwLock` for more details.