feat: Add randomized mutation batching
The randomized batching algorithm works by attempting to put every mutation in a random mutant, creating new mutants if the attempt failed. The random seed can be specified, as well as the number of attempts the algorithm makes to place each mutation into a random mutant.

The randomized batching algorithm is intended to be a lower-end baseline to compare other batching algorithms against.