feat: Replace parallel AST-HIR visiting from mutation engine with maps
The mutation engine now does not directly rely on the parallel visiting of the AST and HIR. Instead, for each function body, a mapping can be built (which still uses the parallel visitor under the hood), which maps AST nodes to their corresponding HIR nodes and vice versa. This map is then used in what is now just a regular AST visit. This has the advantage of not needing a special visitor, and not having to do any bookkeping of the paired up nodes, both in the mutation collector, and the mutation operators. Mutation operators still have access to the same HIR information, but through the body resolutions mappings instead. This also has the advantage of making the parallel visit more testable, as the resulting mapings are more inspectable.