feat: Make MutLoc reference based, and only use it with operators
MutLoc was previously used both for passing location information to operators, and stored in Mut for the created mutations. Because of this latter use case, it had to own its data. However, this is highly inefficient, as all instances contained large amounts of the AST (at least an entire function and many times large subtrees of it alongside it as well). Even disregarding the cloned parts in the heap, MutLoc used to take up 296 bytes. With this change, it now only takes up 24 bytes, with no additional heap memory used. This is done by clarifying the use of MutLoc as a container of short-lived AST location references that get passed to mutation operators, and not as a storage for AST subtrees that can be stored along with Mut instances. This latter use case is not even needed, as only the location span was used after mutation generation.