feat: Add simulated annealing mutation batching
Simulated annealing is a probabilistic search method for approximating the global optimum state. For mutation batching, the search states are batch configurations, and our energy function to be minimized is the number of batches. Sufficiently close neighbouring states are picked by moving mutations from its original batch to another compatible one, one at a time. The acceptance probability function used is the standard one in literature. Currently, we use a linear temperature function.

This simple implementation of simulated annealing seems to perform almost on par with greedy batching almost all the time. This makes it, and other similar search methods, very promising for use in mutation batching.