fix: Parenthesize ambiguous cast expressions in generated code
When cast expressions are placed on the left-hand side of a less-than comparison expression, the right-hand side is ambiguous and is interpreted as an invalid const type argument for the type the left hand side is being cast to (e.g. `len as u32 < 8` is interpreted as `len as u32<8>`). To avoid this, we simply parenthesize every cast expression.