feat: Replace print subcommands with --print argument for many outputs
Previously, we were limited to writing out only one piece of information per invocation of mutest because the printing options were individual subcommands. To support printing multiple pieces of information in a single run of mutest, printing is now done through the --print argument, which accepts a comma-delimited list of information to print. This can be used with the existing build and run subcommands, and the new unified print subcommand, which runs mutest up until every --print option has been printed. This also means that multiple `//@ print-*` UI test directives can be combined with each other, and with `//@ build` and `//@ run`.