fix: Avoid linker errors on some platforms related to f128 cast
The current rustc support for f16 and f128 is highly incomplete. On Apple Silicon the inclusion of an f128 to f64 cast results in the linker error: `Undefined symbols for architecture arm64: "___trunctfdf2"`. The same cast fails on Windows under MSVC with the equivalent linker error: `unresolved external symbol __trunctfdf2 referenced in function`.

To resolve this, and to get mutest-rs building on these platforms again, instead of the cast which is handled by LLVM, we now use the manual reimplementation in rustc_apfloat, which is already used in the compiler to represent these float values.