fix(call_ignore): Mutate calls which return `T: Default` value
We were using an empty, dummy parameter environment for checking if the type of the value of a function call implemented Default, which meant that a `T: Default` was never considered, and function calls which had a generic `T: Default` return type would never get mutated. This is now fixed as we pass the function's parameter environment to the type checker.