feat: Refer to items in inherent and trait impls when sanitizing paths
Previously, we were not able to find visible paths to items in impls, as they are not named module children. The code around these cases have been significantly reworked and clarified. Now, code handling def paths normalizes def paths containing inherent impls to use the path to the implementing type as a base, and reject def paths containing references to trait impls (but not trait assoc items). The idea is that items in trait impls are resolved by getting the path to the trait assoc item, and adding the necessary `Self` qualifier afterwards.