fix: Ensure that we transplant lifetimes according to their local bounds
Previously, when transplanting local lifetime trait bounds onto trait subpaths in qualified self positions, we did not use the local lifetime bounds in the case of a trait impl, but the trait's definition-site bounds instead. This meant that if the lifetime bounds did not have the same names in the impl as in the trait def, or more subtly, if the trait was defined in a different syntax context (e.g. outside of macros) from the impl (e.g. inside a macro invocation), then the extracted lifetime bounds would be incorrect.

To fix this, we now instantiate the trait predicates with the impl's trait arguments. We also extend the tests to assert this behavior.