feat: Verify that AST nodes are mapped to their corresponding HIR nodes
Previously, we were unable to ensure in our AST-HIR mapping tests that all mappable AST nodes were mapped to their corresponding nodes.

To solve this, and the need for future testing-related verification, we add an eternally unstable `--Zverify` option. The first of these is the `--Zverify=ast_lowering` option, usable in tests using the `//@ verify: ast_lowering` directive, which checks every mappable AST node to see if its id is present in the AST-HIR mapping, and emits warnings for nodes that are missing their mapping to an HIR node.