fix: Resolve paths using node res instead of the res of the last segment
This fixes an issue with primitive types, whose path has different resolution if other definitions with the same name are present in the same scope. The last path segment res points to the other item while the node res points to the primitive type. In all other cases, these two resolutions line up.