feat: Print qualified paths for alias projection types
We used to only print the definition itself, not the actual type that was being projected. There was a workaround in hygiene code around this which added the necessary self type qualification on top, but this only worked for top-level types generated in this context.

This behavior is now implemented directly into `AstTyPrinter`, which means that such aliases are resolved without any ambiguities even in nested type argument positions.