feat: Sanitize use imports
This commit has gone through so many iterations with widly different approaches for how to correctly generate sanitized import paths while retaining the original resolutions across all three namespaces. In the end, the solution implemented here, reusing the existing logic for fundamental visible path generation for modules, and module child-based logic for the item leaf nodes, seems simple and robust.

The only hack in this implemenation is related to the unsettled resolution rules around relative use import paths in macros 2.0 expansion contexts, which are supposed to be resolved against the macro's defining scope. This behaviour is implemented, however, currently rustc only accepts such relative paths inside the top-level of macros 2.0 expansions, and rejects them in nested modules within the expansion. This behaviour is undocumented, but is tested for, and this implementation also follows these rules.