feat: Name local items directly when sanitizing import paths
This does not change the resolution of import paths (in all but one case, see below), but makes the paths more similar to hand-written ones, and more importantly, removes the need for visible path generation in the common case, leading to much less work.

This also fixes a specific issue around local macro_rules! items, specifically when the path used when reexporting them, as rustc seems to be quite particular in how you must name such items for the path to resolve. This local macro_rules reexport is done by alacritty for example.