feat: Rework and complete support for sanitizing paths to inherent impls
While we had partial support for generating visible paths to assoc items in inherent impls, this only worked with a limited subset of implementee types and had some workarounds for others (e.g. slices).

With this change, `DefPath` can now represent type-rooted paths directly, representing references to inherent impls on the subject type. This solution supports all implementee types, no matter how complex they are, and reuses existing mechanisms for appending the correct qualified self type to point to the right inherent impls. This change also adds explicit checks for when a type-relative path would be generated in a context which does not accept qualified paths. With this change, we also enumerate all the possible path root kinds a `DefPath` can have, and handle them with a dedicated root type everywhere.