feat: Stop writing infer generics for new path segments when sanitizing
This mechanism was used when we did not ensure that the final, "item" generics from the original path were copied over in all cases. In those cases, adding infer generics in their place made things work. However, since these infer generics were only ever actually needed for the final, "item" path segment, which is now handled properly, and because they caused issues elsewhere, we are removing this mechanism entirely.

The test is repurposed, but may need to be adjusted later for it to make more sense.