fix: Consider std_detect implicit crate dependency when sanitizing paths
The std_detect crate is a dependency of the std crate and the various is_arch_feature_detected! macros generate paths pointing into this crate that are not visible from the std crate and are only visible from std_detect directly. Since this crate is not part of the extern prelude, lookups for the crate's children failed when finding visible paths pointing to them.

To solve this issue, we add this implicit dependency as an extern crate to the base compiler options used for both analysis/codegen and compilation. We already did this for the compilation pass, but we need it for the analysis/codegen pass as well for successful resolution. This change simply moves the extern declaration into a common place.