feat: Resolve mod paths in imports the same way for mods and child items
For import paths, we were generating potentially different paths for a module versus one of its items, in the same scope. The main reason was that we did not differentiate between generating a fully valid module path, and a parent module path stub. These, for the sake of validity, would not necessarily match up with each other.

Now we differentiate between the two cases, so that we can generate the same base module paths in such cases, while keeping the final paths valid in all cases.