feat: Compact generated code for subst maps to improve compile times
Previously, we used to generate large struct literals to represent substitution maps in the generated code, resulting in huge bloat in the generated source code, which ultimately resulted in excessive compile times.

With this change, we now use a significantly more compact representation, a list of only the active substitution metadata entries, to represent these sparse substitution maps in generated code. At runtime, the substitution maps are represented as fixed length arrays, which are equivalent to the previously used generated struct definitions. The conversion between the compact source code form and the lookup-efficient in-memory form of the substitution maps is handled by a generic const function in the runtime crate. This is evaluated by the compiler much faster than the large struct literals we were using before.

To put things into perspective, I measured the amount of generated source code and compile times before and after this change on the image crate with 4662 mutations. This change reduces the amount of generated source code from 871.79 MiB (11,602,353 SLoC) down to just 12.80 MiB (233,830 SLoC), a reduction of about 98.53% in terms of MiB, and a reduction of about 97.98% in terms of SLoC. This results in compile times going from 197.19s down to just 13.75s, which is 14.34 times faster, and takes 93.03% less time. The size of the resulting binaries is also reduced slightly from 181 MiB down to 169 MiB, by about 6.63%.

These are massive improvements that improve usability significantly, without introducing any changes in runtime characteristics or performance. These new, massively improved compile times also clearly show that our meta-mutant approach benefits mutation testing in Rust greatly! :-)