feat(call_graph): Collect extern crate function calls and their callees
While we only target crate-local functions in the call graph, we need to construct the full call graph, including the calls made to extern crate functions and their callees, because these functions could make calls back to the local crate. This can happen for example with a generic extern function that takes a type implementing a trait. If a type with a local crate impl is passed in, then that extern function can make calls back to the local crate.

With this change, such calls are accounted for, which corrects test-target discovery and mutation batching mistakes/incompleteness present with hashbrown and other crates.