test(call_graph): Add MIR-less call to harden extern crate call test
Extern function callees are discovered through their MIR metadata entries, which are not present for all items. For example, private, non-generic functions do not have their MIRs saved into the crate metadata for efficiency. This is fine, because such cases cannot call back to the local crate, which is what we are ultimately looking for.

This test case addition ensures that we handle the lack of MIR availability during call graph construction gracefully.