fix: Avoid targetting skipped fns containing non-skipped closures
This is the result of a very intricate bug in `ast_lowering::find_def_in_ast` which currently incorrectly resolves closure defs to their containing function's AST item. This resulted in an apparent issue in mutation generation, where targeted closures mutated their containing function instead, which undid any effects of annotating a function with the `#[mutest::skip]` attribute.

This change does not fix the bug in `ast_lowering::find_def_in_ast. Instead, it adds an assertion to the function which gets triggered when such an AST resolution mismatch happens, for future reference. For now, since we do not mutate closure bodies, we simply do not consider them from targeting during call graph construction. Eventually, when we want to support mutating closure bodies, this will have to be undone, and the bug in `ast_lowering::find_def_in_ast` will need to be fixed.