fix(call_graph): Get unsafe block contexts when resolving MIR callees
This fixes a regression that resulted from us moving to MIR-based callee resolution in the local crate. The previous AST-based callee resolution collected whether the calls were made from unsafe blocks or not. This information is not directly present in the MIR, however, we can match MIR call spans up with the spans of unsafe blocks in the HIR (including careful consideration for macro expansions) to achieve the same effect.