feat(runtime): Build in-memory mutation subsumption matrix
Calculating the individual mutation subsumptions can be relatively expensive, especially for large mutation subsumption matrices with lots of underlying test case results to compare.

To alleviate any unnecessary computation, this change makes it so that we now pre-compute the entire subsumption matrix before printing. Since the matrix is mirrored across its diagonal axis (aka. a triangular matrix), we are now able to do only half of the subsumption computations by calculating the entries in the upper triangular part of the matrix, and inserting the reverse entries in the lower triangular part of the matrix.