feat(call_graph): Disable inlining for completeness with optimizations
When optimizations are enabled, rustc performs inlining on the MIR bodies, which our analysis uses to construct the definition-level call graph. This inlining thus affects the output of call graph construction, and we can miss inlined definitions in optimized builds.

To ensure that MIR bodies, and this the resulting call graph, is complete and matches between optimized and unoptimized builds, we disable all MIR optimizations during code analysis. It is worth noting that this does not affect the optimizations of the final output, as the analysis MIR is never used for any compilation, only for constructing the meta-mutant AST.