feat: Make mutation batches optional in driver's internal representation
Regardless of whether mutation batching was enabled, mutations were always moved into `Mutant` grouping structs in the driver's internal representation. This causes unnecessary friction with code that does not deal with mutation batches. It also impedes our ability to introduce non-batched mutation evaluation models. To solve these issues, this change ensures that mutation ownership is always with the mutations collection regardless of mutation evaluation model, makes mutation batch descriptors reference based, and generally makes mutation batches optional throughout the driver code.

This change removes almost all mentions of the term "mutant" in the driver code, referring to them as "mutation batches" instead. For consistency, the `--print=mutants` flag has also been renamed to `--print=mutations`, which also better describes the print output. Batching groupings are only printed when mutation batching is enabled, and are ommitted otherwise in favor of a simpler output.

This change deliberately does not touch the generated code, and a similar change in the runtime and corresponding generated code is coming as a follow-up commit to this.