feat(driver): Embed runtime paths and option for embedding files
Previously, we relied on the user providing the build directory of their local mutest-rs checkout using the MUTEST_SEARCH_PATH environment variable. This was both a bad user experience, but also the implementation of this was flaky at best, incorrectly relying on slightly incorrect output from `cargo --build-plan`.

Instead, we now have two approaches. For development, we now simply embed the full file paths to the relevant build directories directly into the mutest-driver binary, defaulting to these paths unless MUTEST_SEARCH_PATH is specified, and for "release" builds, we now have the build option `--features embed-runtime` to embed the runtime `rlib` and dependency files by directly packing them into the mutest-driver binary, and unpacking them into `target/mutest/mutest_deps` during execution.

This change also removes the invalid build script of mutest-runtime, and the direct dependency of mutest-driver on mutest-runtime. This means that mutest-runtime needs to be built using a manual Cargo invocation, although this was effectively the case before as well. This is also ensured by a check in mutest-driver's build script.