feat: Add support for mutating for and running against integration tests
Previously, we did not support running integration tests at all, and simply failed in a way that was unaccounted for if any integration tests were targeted by the Cargo invocation.

This change makes it so that we now handle integration tests through the compiler-side, by plugging into the regular underlying `cargo test` rustc invocations the best we can. For regular unit tests, this changes nothing, and the behavior remains the same as before. For integration tests, we use the fact that Cargo builds a regular library target for integration tests to link against, by building a generic meta-mutant in this pass, using the crate's public interface functions as entry points, rather than the crate's test cases. Then, for each integration test, we use the test case call graph of the integration test to target not local mutable functions, but function calls into the extern library target. Integration tests have their own generated harness entry point, which drives the meta-mutant of the extern crate with the tests of the integration test, alongside some additional metadata needed to connect the reachability analysis across the crate boundary between the two seperate crates.

The one major downside to this approach is that we cannot support the generic function entry points of the meta-mutant crate, as we cannot generate the fully monomorphized call graph for the meta-mutant crate without knowing all of the possible external test case entry points, and their monomorphized call graphs up front. However, this is a big step forward covering many use cases, and we can adjust our compilation and crate analysis strategy later to remove this limitation in the near future.

This support for integration tests also implies the need for adding Cargo project-based tests, which will come in a follow up commit.