fix: Remove analysis-generated main and mutate no harness test targets
The previous change broke bin crates due to duplicate main definitions. We now solve this issue in a much more resilient way, by simply removing the test-generated entry point in the analysis pass, as it will be regenerated during the compilation pass anyway.

Both for writing a test for this fix, and for better compatibility with custom test harnesses (e.g. embedded-test), we now treat `--cfg=test` invocations as test targets too, not just `--test` invocations. This corresponds with the `harness = false` option for `[[test]]` targets in `Cargo.toml` package manifests, which is commonly required by custom test harnesses to supress libtest injection.