feat(embedded): Override Cargo test bin runner with runtime host driver
This change automatically configures the runner program for the generated meta-mutant binaries to the embedded mutation runtime host driver, overriding any existing configurations for generic flashing and monitoring tools, such as probe-rs or espflash. The embedded mutation runtime host driver is responsible for flashing the generated firmware image onto a suitable target (potentially multiple in the future), and driving the mutation evaluation in cooperation with the generated target firmware.